Changed encryption script to not take user input. Preparing for being called from HTML
